вынесены интерфейсы из accessToken.guard.ts в отдельный файл в auth/types.ts
