+++ src/auth/guards/refreshToken.guard.ts
@@ -5,7 +5,7 @@ import {
   UnauthorizedException,
 } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
-import { AuthRequest, JwtPayload } from './accessToken.guard';
+import { JwtPayload, AuthRequest } from '../types'
 import { ConfigService } from '@nestjs/config';
 
 @Injectable()
